Replace userbind with a more practical LD_PRELOAD implementation
The configuration needed to provide each user with their own network
namespace is complex and intrusive on the host. An alternative, less-
sophisticated route uses LD_PRELOAD to intercept privileged bind()
calls.

To allow a program to bind to privileged ports, include libwrapbind.so
in LD_PRELOAD: a helper script 'wrapbind' is provided to do this. The
libwrapbind.so library wraps bind(), routing privileged attempts to a
setuid helper binary. After checking the address is permitted for the
invoking user in /etc/addresses, the helper performs the bind on behalf
of the unprivileged process.

Alternatively, build userbind into binaries by including -luserbind in
LDLIBS (to link against libuserbind.a) and adding -Wl,--wrap=bind to
LDFLAGS. It is also possible to call the helper 'userbind' directly: it
expects the socket to bind on stdin and takes the address and port as
arguments.

The /etc/addresses configuration file consists of lines of the format

  USER:ADDRESS[/MASK]

where USER is a username or numeric uid, ADDRESS is a permitted IPv4
address in dotted-quad notation or IPv6 address in canonical notation,
and MASK is optional a decimal netmask to allow an address range to be
specified in CIDR format. There can be multiple lines per user, and
0.0.0.0/0 or ::/0 can be listed to allow a user to bind to all IPv4 or
IPv6 addresses respectively.