Use data.new as a temporary link name
data.tmp is used as a temporary file by microdns when it compiles data.cdb
from data. Although dnsconfig wouldn't normally be configured to edit
data in the live microdns server directory, choose a different temporary
link name for safety.

Don't attempt to unlink data.new before linking our O_TMPFILE fd to it.
A stale data.new link should never be present when we have the lock,
so it's better to detect this and error out than silently cover it up.

Sadly, linkat() doesn't support a force/overwrite flag to allow the
rename() step to be skipped altogether.